Extract tests from semantic/assembly/mod.rs
Extracted the massive inline `mod tests` block from `src/semantic/assembly/mod.rs` into a new file `src/semantic/assembly/tests.rs` to alleviate file bloat.
